Extract Operation to its own file
